email validation, shifts clients creation
