+++ api/app.py
@@ -1,11 +1,18 @@
 
+from crypt import methods
+from wsgiref.validate import validator
 from sqlalchemy_utils import ChoiceType
 from flask_sqlalchemy import SQLAlchemy
-from flask import Flask, request, json
+from flask import Flask, request, json, render_template
 from flask_restx import Resource, fields, Api
 from flask_cors import CORS
 from loggingFile import *
 from utils.sendEmail import *
+from flask_wtf import FlaskForm
+from wtforms import StringField, PasswordField, BooleanField
+from wtforms.validators import InputRequired,Email,Length
+
+
 app = Flask(__name__)
 api = Api(app)
 
@@ -18,6 +25,11 @@ TYPES = [
     ('available', 'Available'),
     ('not available', 'Not available')
 ]
+###########FORMS#############
+
+class LoginForm(FlaskForm):
+    email = Email
+    password = PasswordField('password',validators = [InputRequired(), Length(min=8,max=80)])
 
 
 ########### MODELS #############
@@ -65,7 +77,7 @@ class Company(db.Model):
 class Client(db.Model):
     __tablename__ = 'client'
     id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String(80), unique=True, nullable=False)
+    name = db.Column(db.String(80), nullable=False)
     email_address = db.Column(db.String(120), unique=True, nullable=False)
 
     def _repr__(self):
@@ -87,12 +99,40 @@ user_serializer = api.model('Model', {
 
 })
 
-# ########### ROUTES #############
-@api.route('/hello')
-class HelloWorld(Resource):
-    def get(self):
-        return {'hello': 'world'}
+client_serializer = api.model('Model', {
+    'id': fields.Integer(readOnly=True, description='The unique identifier of client'),
+    'name': fields.String(required=True, description='The name of client'),
+    'email_address': fields.String(required=True, description='The email address of user'),
+})
 
+schedule_serializer = api.model('Model', {
+    'id': fields.Integer(readOnly=True, description='The unique identifier of schedule'),
+    'user_id': fields.Integer(required=True, description='The unique identifier of cleaner'),
+    'client_id': fields.Integer(required=True, description='The unique identifier of client'),
+    'date_range': fields.String(required=True, description='The details of the shift')
+})
+
+# ########### ROUTES #############
+
+#return me all the scheduled shifts in db
+@api.route("/shifts")
+class Shifts(Resource):
+    @api.marshal_with(schedule_serializer, envelope='resource')
+    def get(self):
+        return Schedule.query.all();
+
+@api.route("/client/<string:email>")
+class ClientsByEmail(Resource):
+    @api.marshal_with(client_serializer, envelope='resource')
+    def get(self, email):
+        return Client.query.filter_by(email_address=email).all()
+
+#returns me all the clients in db
+@api.route("/clients")
+class Clients(Resource):
+    @api.marshal_with(client_serializer, envelope='resource')
+    def get(self):
+        return Client.query.all();
 
 @api.route('/companies')
 class Companies(Resource):
@@ -107,13 +147,27 @@ class Users(Resource):
     def get(self, company_id):
         return User.query.filter_by(company_id=company_id).all()
 
-# @api.route('/setAppt', methods={'POST'})
-# class SetAppt(Resource):
-#     def post():
-#         if(request.method == 'POST'):
-#             client = request.form.get("client")
-#             console.log(client)
-#             return {"client":client}
+# client route creating new clients uniquely identified by their email ids
+@app.route("/createClient", methods=['POST'])
+def createClient():
+    if request.method == 'POST':
+        request_data = json.loads(request.data)
+        client = Client(name=request_data.get('name'),email_address=request_data.get('email'))
+        db.session.add(client)
+        db.session.commit()
+        
+        return str(client.id);
+
+#shift route creating new shift schedule using client and user ids and the schedule
+@app.route("/createShift", methods=['POST'])
+def createShift():
+    if request.method == 'POST':
+        request_data = json.loads(request.data)
+        shift = Schedule(user_id=request_data.get('user_id'),client_id=request_data.get('client_id'),date_range=request_data.get('date_range'))
+        db.session.add(shift)
+        db.session.commit()
+        
+        return str(shift.id);
 
 
 @app.route("/setAppt", methods=["POST"])
@@ -122,10 +176,16 @@ def setAppt():
         request_data = json.loads(request.data)
         #print(request_data['content'])
         
+        #create new entitiy in db for the schedule
         #logging shift details in log file
         append_new_line('log.txt', request_data.get('client') + " scheduled " + request_data.get('cleaner') + " from "
         + request_data.get('start') + " for " + request_data.get('hours')+ " hours on " + request_data.get('date'))
         
+
+
+        # db.session.add(Client(name=request_data.get('client').split("@")[0],email_address=request_data.get('client')))
+        # db.session.commit()
+
         #send the confirmation email to the client and the cleaner
         sendConfirmationEmail(request_data.get('client'),request_data.get('cleaner').split(":")[1],
             {
