+++ api/log.txt
@@ -1,21 +1,1 @@
-karanpreetsodhi1997@gmail.com scheduled Cindy Smith: cindy.smith@companyabc.com from 14:00 for 5 hours on 2022-08-24
-karanpreetsodhi1997@gmail.com scheduled Cindy Smith: cindy.smith@companyabc.com from 13:05 for 0 hours on 2022-09-01
-karanpreetsodhi1997@gmail.com scheduled Cindy Smith: cindy.smith@companyabc.com from 14:06 for 1 hours on 2022-08-23
-karanpreetsodhi1996@gmail.com scheduled Cindy Smith: cindy.smith@companyabc.com from 14:20 for 2 hours on 2022-08-22
-karanpreetsodhi1997@gmail.com scheduled Cindy Smith: cindy.smith@companyabc.com from 14:21 for 2 hours on 2022-08-25
-karanpreetsodhi1997@gmail.com scheduled Harry Potter: harry.potter@companyamz.com from 14:21 for 3 hours on 2022-08-31
-karanpreetsodhi1997@gmail.com scheduled Cindy Smith: cindy.smith@companyabc.com from 20:19 for 3 hours on 2022-08-22
-karanpreetsodhi1996@gmail.com scheduled Cindy Smith: cindy.smith@companyabc.com from 14:26 for 4 hours on 2022-08-30
-karanpreetsodhi1996@gmail.com scheduled Cindy Smith: cindy.smith@companyabc.com from 14:26 for 6 hours on 2022-09-01
-karanpreetsodhi1997@gmail.com scheduled Cindy Smith: cindy.smith@companyabc.com from 15:21 for 1 hours on 2022-08-24
-karanpreetsodhi1996@gmail.com scheduled Cindy Smith: cindy.smith@companyabc.com from 20:21 for 5 hours on 2022-08-23
-karanpreetsodhi1997@gmail.com scheduled Cindy Smith: cindy.smith@companyabc.com from 20:37 for 2 hours on 2022-08-24
-karanpreetsodhi1997@gmail.com scheduled Karan Sodhi: karanpreetsodhi1996@gmail.com from 14:50 for 4 hours on 2022-08-23
-karanpreetsodhi1997@gmail.com scheduled Karan Sodhi: karanpreetsodhi1996@gmail.com from 16:50 for 3 hours on 2022-08-26
-karanpreetsodhi1997@gmail.com scheduled Karan Sodhi: karanpreetsodhi1996@gmail.com from 03:00 for 1 hours on 2022-08-23
-karanpreetsodhi1997@gmail.com scheduled Karan Sodhi: karanpreetsodhi1996@gmail.com from 05:30 for 11 hours on 2022-08-23
-karanpreetsodhi1997@gmail.com scheduled Karan Sodhi: karanpreetsodhi1996@gmail.com from 18:26 for 2 hours on 2022-08-24
-karanpreetsodhi1997@gmail.com scheduled Karan Sodhi: karanpreetsodhi1996@gmail.com from 15:00 for 3 hours on 2022-08-23
-karanpreetsodhi1997@gmail.com scheduled Karan Sodhi: karanpreetsodhi1996@gmail.com from 19:37 for 1 hours on 2022-08-24
-karanpreetsodhi1997@gmail.com scheduled Karan Sodhi: karanpreetsodhi1996@gmail.com from 18:00 for 2 hours on 2022-08-24
-karanpreetsodhi1997@gmail.com scheduled Karan Sodhi: karanpreetsodhi1996@gmail.com from 17:48 for 2 hours on 2022-08-22
+karna@gmma.com scheduled Cindy Smith: cindy.smith@companyabc.com : 3 from 06:02 for 2 hours on 2022-08-23
