+++ web/package.json
@@ -7,10 +7,14 @@
     "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^0.27.2",
+    "email-validator": "^2.0.4",
+    "isemail": "^3.2.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-flash-message": "^1.0.8",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
+    "validator": "^13.7.0",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
