+++ web/src/pages/Login.js
@@ -1,0 +1,11 @@
+import React from "react"
+
+const Login = ()=>{
+
+    return(
+        <h3>Login Page</h3>
+    )
+
+}
+
+export default Login
