+++ web/src/pages/NoMatch.js
@@ -1,0 +1,13 @@
+import React from "react";
+
+const NoMatch = () => {
+
+    return (
+        <>
+            <h2>Oops, we cant find the page.</h2>
+        </>
+    )
+
+}
+
+export default NoMatch;
