+++ web/src/pages/Signup.js
@@ -1,0 +1,11 @@
+import React from "react"
+
+const Signup = ()=>{
+
+    return(
+        <h3>Signup Page</h3>
+    )
+
+}
+
+export default Signup
